requirements.txt
spacy
nltk
indic-nlp-library
inltk
pandas
pyyaml
streamlit  # or gradio
